add error handling with 404

+++ main.py
@@ -1,4 +1,4 @@
-from fastapi import FastAPI, Depends
+from fastapi import FastAPI, Depends, HTTPException
 from pydantic import BaseModel
 from sqlalchemy.orm import Session
 from database import SessionLocal, engine
@@ -27,6 +27,13 @@ def hello():
 def get_tasks(db: Session = Depends(get_db)):
     return db.query(models.Task).all()
 
+@app.get("/tasks/{id}")
+def get_task(id: int, db: Session = Depends(get_db)):
+    task = db.query(models.Task).filter(models.Task.id == id).first()
+    if task is None:
+        raise HTTPException(status_code=404, detail="Task not found")
+    return task
+
 @app.post("/tasks")
 def create_task(task: TaskSchema, db: Session = Depends(get_db)):
     db_task = models.Task(title=task.title, done=task.done)
@@ -38,6 +45,8 @@ def create_task(task: TaskSchema, db: Session = Depends(get_db)):
 @app.delete("/tasks/{id}")
 def delete_task(id: int, db: Session = Depends(get_db)):
     task = db.query(models.Task).filter(models.Task.id == id).first()
+    if task is None:
+        raise HTTPException(status_code=404, detail="Task not found")
     db.delete(task)
     db.commit()
     return {"message": f"Task {id} deleted"}
@@ -45,6 +54,8 @@ def delete_task(id: int, db: Session = Depends(get_db)):
 @app.put("/tasks/{id}")
 def update_task(id: int, updated_task: TaskSchema, db: Session = Depends(get_db)):
     task = db.query(models.Task).filter(models.Task.id == id).first()
+    if task is None:
+        raise HTTPException(status_code= 404, detail="Task not found")
     task.title = updated_task.title
     task.done = updated_task.done
     db.commit()
